Breaks up watchdogs into subpackage
The udev_watchdog has been broken up in to a BaseWatchdog class
in a base.py module and a Watchdog class in a linux.py module.

In future will have a windows.py watchdog, with the one loaded
depending on the operating system.
