+++ src/automakemkv/watchdogs/__init__.py
@@ -1,0 +1,12 @@
+"""
+Utilities for ripping titles
+
+"""
+
+import signal
+from threading import Event
+
+RUNNING = Event()
+
+signal.signal(signal.SIGINT, lambda *args: RUNNING.set())
+signal.signal(signal.SIGTERM, lambda *args: RUNNING.set())
